Remove `full_bounds`
Original intention was that `bounds` indicate the bounds from which factors values are sampled from, and `full_bounds` indicate the "true" bounds to display to users (given the flooring trick to transform from continuous to discrete).

This is largely unused though and a possible source of confusion, so am removing it until there's a need for it.