Increase minimum number of scenarios required to run in parallel
In usual circumstances, spinning up workers should take ~20-30 seconds.

In some cases, however, it may take 3 - 5 minutes.
Unsure what the cause is but it seems sporadic, and possibly tied to what is running in the background.

Increasing the minimum number of scenarios ensures it's at least worthwhile using parallel processing in the majority of cases.