Fix clusters plot in wrong order
Previously we were using the index of unique(clusters) array to find clusters colors, plot clusters and render the legend. But depending on the order they appear in clusters vector this could led to mistakes, like cluster number 2 being the first in this vector (meaning it has index 1). That was fixed in a previous commit but the legend wasn't.