Replace usage of max_area (k) with 1.0 as cover now already proportional to k
Replace max_cover variable with max_area variable

Previously max_cover was used a mcda variable input as cover was relative to total area so k was needed to work out total space available for coral. As cover is now relative to k, this is no longer needed, so the variable can be removed from the mcda struct. However, area of cover is still needed for the shading matrix.