Avoid assuming where the RCP column is
The column indicating the RCP was originally the first entry but can be the first or last column depending on how/who the scenario spec was generated (and what version of ADRIA).

Remove the assumption that the RCP column is before any other discrete-valued parameter.

Fixed a bug where values set as constants by the user were being floored as well.