Remove return type specification
We've switched to defining the Distribution type instead of the string which indicates what the Distribution is.

Unfortunately, the API from Distributions.jl is not fully stable yet, and the Distributions subtype a mix of different parent types (DataType, UnionAll, etc).

The simple fix here is to simply allow any type to be returned. This is not good for performance or stability, but good enough for now.