Handle change from `TP_data` -> `conn`
TP_data (transition probability) implies the rows should sum to one (source to sink).

Changing to `conn` (for connectivity) which indicates the proportion of larvae that survive and arrive/settle on a sink location. The remaining proportion indicates larvae "lost" out to sea, hence there is no expectation for rows to sum to 1.0