Remove previous workaround
GeoMakie could not plot ArchGDAL geometry types directly, and so a hacky workaround was implemented where the viz module would store a GeoJSON copy of the feature geometries.

This would lead to issues where, for example, multiple domains could be loaded.

This set of changes replaces that workaround with one that handles the geometry objects, removing the need for an on-disk GeoJSON copy to be created.