Fix docs of two analysis blocks
When running the documentation of the time series clustering analysis an
error was returned in the plot. The error happens because the code is
assigning an array as a value to a Dict{Symbol,Bool}. This commit
changes the opts dictionary to define it as a Dict{Symbol,Any}.

In the documentation of the regional sensitivity analysis, when running
the preparation to run the rsa analysis it was returning an error. The
error happens because it is applying dropdims to a float. The float came
from a mean applied to an YAXArray. This commit changes the mean to
apply the mean over one dimension of the YAXArray.