Minor fix of docs of load_domain
The function load_domain always need to have as input the path to the
domain file. Most of the uses of load_domain have inside the parenthesis
"path to domain data package". To keep consistency, this commit adds
this text to two calls of the functions that did not had an input.