Temporary fix on rank_by_index
Running select_locations was causing this error:
ERROR: UndefVarError: `_dp` not defined in local scope
Suggestion: check for an assignment to a local variable that shadows
a global of the same name.

As a temporary solution this commits rename _dp to dp, to avoid creating
a new variable inside an if block.