Adds sample_options method for pathway diversity analysis
This commit adds a new sample method where all parameters are fixed and
only option_ts changes. option_ts is a vector of symbols that defines
the option used in each timestamp. This method will be used in the
pathway diversity analysis. It also modifies the scenario file to allow
receiving this new type of input.