Make cluster_diversity analysis use CB_CALIB_GROUPS
When looking to the RME or ReefMod datasets cluster_id has an unique
identifier that does not allow the cluster_diversity analysis (the
result is a vector of zeros). To run the pathway diversity analysis it
temporarily changes the code to use CB_CALIB_GROUPS.
It also change the direction of the seed_geographic_separation from
minimum to maximum, as in this simulation we always want to maximise
the geographic separation.